refactor: refactoring the return endpoint of a supplier's products and a category's products
